requirements.txt
Flask
flask-cors
scikit-learn
pandas
numpy
joblib
python-dotenv
requests
gunicorn